misc/HtmlTestSuite/results/test-0814.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Text: A sequence of collapsible spaces at the end of a line is removed</title>
<link href="https://drafts.csswg.org/css-text-3/#white-space-phase-2" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="/fonts/ahem.css" rel="stylesheet" type="text/css"/>
<style>
#container &gt; div {
  display: inline-block;
  -font-family: Ahem;
  font-size: 10px;
}
</style>
</head><body>
<div id="log"/>
<div id="container">
  <div>1111</div>
  <div>1111 </div>
  <div> 1111</div>
  <div> 1111 </div>
  <div>1111<br/></div>
  <div>1111<br/> </div>
  <div>1111 <br/></div>
  <div>1111 <br/> </div>
</div>
<script>
(function () {
  const epsilon = 1;
  let elements = Array.from(container.children);
  let reference = elements[0];
  let reference_width = reference.offsetWidth;
  for (let element of elements) {
    test(() =&gt; {
      assert_approx_equals(element.offsetWidth, reference_width, epsilon);
    }, escapeSpaces(element.innerHTML));
  }
})();

function escapeSpaces(text) {
  return text
    .replace(/ /g, '&amp;#x20;')
    .replace(/\n/g, '&amp;#x0A;')
}
</script>
</body>
</html>
